refactor useFunction and useMetadata (#143)
